verify value scenario
